refactor: 결과값을 책임지는 클래스명 및 위치 변경 ResultDto->Result
