+++ src/main/java/domain/Result.java
@@ -1,13 +1,10 @@
-package dto;
+package domain;
 
-import domain.Lotto;
-import domain.LottoNumber;
-import domain.RankPrize;
 import java.util.Arrays;
 import java.util.EnumMap;
 import java.util.List;
 
-public class ResultDto {
+public class Result {
 
     private static final int RESULT_INIT_NUMBER = 0;
     private static final int INIT_WIN_PRICE = 0;
@@ -19,7 +16,7 @@ public class ResultDto {
             .forEach(rank -> resultDto.put(rank, RESULT_INIT_NUMBER));
     }
 
-    public ResultDto(final List<Lotto> issuedLotto, final Lotto lastWinLotto, final LottoNumber bonusNumber) {
+    public Result(final List<Lotto> issuedLotto, final Lotto lastWinLotto, final LottoNumber bonusNumber) {
         for (Lotto lotto : issuedLotto) {
             final RankPrize rankPrize = RankPrize.of(lotto.compare(lastWinLotto), lotto.isContainNumber(bonusNumber));
             resultDto.put(rankPrize, resultDto.getOrDefault(rankPrize, 0) + 1);
